Added fix to problems with generating notes when deploying app to heroku
Section describing process of connecting mongodb to the notes application backup relies on the localhost development environment. On local instance everything will work as database connection info is stored in the .env file. However when  backend + frontend is deployed to heroku list of notes will be empty. 
Heroku  logs are pointing to following error: 

error connecting to MongoDB: The `uri` parameter to `openUri()` must be a string, got "undefined". Make sure the first parameter to `mongoose.connect()` or `mongoose.createConnection()` is a string.
After adding env variable  manually to project setup everything works as expected on heroku. 

Suggestion: maybe this idea could be extented further and/or screenshots could be added?